Create mrb_http2_config.{c,h}
